Refactor contexts folder structure
