+++ src/Contexts/Backoffice/Courses/application/Create/CreateBackofficeCourseOnCourseCreated.ts
@@ -1,6 +1,6 @@
-import { CourseCreatedDomainEvent } from '../../../Mooc/Courses/domain/CourseCreatedDomainEvent';
-import { DomainEventClass } from '../../../Shared/domain/DomainEvent';
-import { DomainEventSubscriber } from '../../../Shared/domain/DomainEventSubscriber';
+import { CourseCreatedDomainEvent } from '../../../../Mooc/Courses/domain/CourseCreatedDomainEvent';
+import { DomainEventClass } from '../../../../Shared/domain/DomainEvent';
+import { DomainEventSubscriber } from '../../../../Shared/domain/DomainEventSubscriber';
 import { BackofficeCourseCreator } from './BackofficeCourseCreator';
 
 export class CreateBackofficeCourseOnCourseCreated implements DomainEventSubscriber<CourseCreatedDomainEvent> {
