--- src/Contexts/Backoffice/SearchAll/application/SearchAllCoursesQuery.ts
@@ -1,3 +1,0 @@
-import { Query } from '../../../Shared/domain/Query';
-
-export class SearchAllCoursesQuery implements Query {}
