+++ src/apps/backoffice/backend/config/dependency-injection/application_test.yaml
@@ -4,7 +4,8 @@ imports:
 services:
   Backoffice.Backend.EnvironmentArranger:
     class: ../../../../../../tests/Contexts/Shared/infrastructure/mongo/MongoEnvironmentArranger
-    arguments: ['@Shared.ConnectionManager']
+    arguments: ['@Shared.MongoConnectionManager']
+
   Backoffice.Backend.ElasticEnvironmentArranger:
     class: ../../../../../../tests/Contexts/Shared/infrastructure/elastic/ElasticEnvironmentArranger
-    arguments: ['@Shared.ConnectionManagerElastic']
+    arguments: ['@Shared.ElasticConnectionManager']
