+++ tests/Contexts/Mooc/Courses/domain/CourseMother.ts
@@ -1,12 +1,11 @@
-import { CourseId } from '../../../../../src/Contexts/Mooc/Shared/domain/Courses/CourseId';
-import { CourseName } from '../../../../../src/Contexts/Mooc/Courses/domain/CourseName';
-import { CourseDuration } from '../../../../../src/Contexts/Mooc/Courses/domain/CourseDuration';
-import { Course } from '../../../../../src/Contexts/Mooc/Courses/domain/Course';
-import { CreateCourseRequest } from '../../../../../src/Contexts/Mooc/Courses/application/CreateCourseRequest';
-import { CourseIdMother } from '../../Shared/domain/Courses/CourseIdMother';
-import { CourseNameMother } from './CourseNameMother';
-import { CourseDurationMother } from './CourseDurationMother';
 import { CreateCourseCommand } from '../../../../../src/Contexts/Mooc/Courses/application/CreateCourseCommand';
+import { Course } from '../../../../../src/Contexts/Mooc/Courses/domain/Course';
+import { CourseDuration } from '../../../../../src/Contexts/Mooc/Courses/domain/CourseDuration';
+import { CourseName } from '../../../../../src/Contexts/Mooc/Courses/domain/CourseName';
+import { CourseId } from '../../../../../src/Contexts/Mooc/Shared/domain/Courses/CourseId';
+import { CourseIdMother } from '../../Shared/domain/Courses/CourseIdMother';
+import { CourseDurationMother } from './CourseDurationMother';
+import { CourseNameMother } from './CourseNameMother';
 
 export class CourseMother {
   static create(id: CourseId, name: CourseName, duration: CourseDuration): Course {
